pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "spirit-anime"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask-restful = "^0.3.10"
requests = "^2.31.0"
pytest = "^8.1.1"
flask-sqlalchemy = "^3.1.1"

[tool.pytest.ini_options]
pythonpath = ["src"]
